Added more functions to the game
I added a new function to have the player make next move on the board.
I've also changed the code around a bit so it functions better. I had
forgotten to create a global variable for the players and originally had
them as local variables. I have tested it and it works as intended.