Added a new feature and new variable to track game
I added a new method within the factory function. I added a getTurn()
and setTurn() method to track the turn of the game. This is so in the
addition when I add the AI of the game, I can have the program go
automatically after the player has gone.

The other feature goes hand in hand with the added method. I added a
checkTurn() method to check which player is going. This is so when the
player or the computer goes, the program will know which player will go
next and will change the turn accordingly.