A lot of headway so far
I was able to create a loop for the DOM to create game spaces and allows
the player and computer to put their mark on the board. However, right
now it isn't clickable. I've also connected the game wins to the DOM.
Also, I was able to get the gameReset() function to work and it clears
the game board. However, I haven't connected it to the DOM/ the reset
button yet. I've also did some CSS work as well to make it look a bit
better. I used mostly flexbox for this so far.