More updates
I was able to tweak the css more and now I have a tic tac toe gameboard.
Also, made the background black, the text white, and the board grey. I
ensured that the board was centered on the page.

As for the js responsiveness, I was able to get the player to click on
the so when the player clicks on the board, the x or o will appear on it
depending on which player's turn it is. Also, whoever wins, the
gameboard highlights the winning moves. However, now the computerMoves()
function is not working. I am not sure why. I will have to look into it.
Also, something else I want to look at is the displaying the winner,
draw, and game over messages.