Nearly done and just need to clean up
I pretty much got the complete functionality of the game working. A few
things I completed was changed the CSS of the game to show which sides
are the players. I also added a new game button that will completely
reset the game.

For the js file, I changed the code so whoever wins the game will be
displayed on the screen. Also, whoever wins the game will go last when
resetting the game. Also, worked out some bugs with the game.

Next I want to clean up the code and make it look better.